+++ docs/changelog.rst
@@ -2,6 +2,18 @@
 Changelog
 *********
 
+v2.0.4 (UNRELEASED)
+===================
+
+Bug fix release.
+
+- It has been observed that libspotify will reuse cached proxy settings from
+  previous sessions if the proxy fields on the ``sp_config`` struct are set to
+  ``NULL``. When the ``sp_config`` fields are set to an empty string, the
+  cached settings are updated. When attributes on :class:`spotify.Config` are
+  set to :class:`None`, we now set the fields on ``sp_config`` to empty strings
+  instead of ``NULL``.
+
 
 v2.0.3 (2015-09-05)
 ===================